breaking cli up into individal files (like fastapi routers) and adding initial localization entrypoints
