+++ analysis/cli/__init__.py
@@ -1,0 +1,4 @@
+from analysis.cli.cli import cli
+from analysis.cli.localization import localization
+
+__all__ = ["cli", "localization"]
