+++ analysis/features/metrics/localization_metrics.py
@@ -1,7 +1,9 @@
 from __future__ import annotations
 
 from pydantic import BaseModel
-from analysis.models.patch import Location, Patch, ScopeKind
+from analysis.models.patch import Patch
+from analysis.models.localization import Location, ScopeKind
+
 
 class LocalizationMetrics(BaseModel):
     file_match: bool
@@ -35,17 +37,72 @@ class LocalizationMetrics(BaseModel):
     """
 
     @staticmethod
-    def localization_of_kind(generated_locations: list[Location], gold_locations: list[Location], kind: ScopeKind) -> tuple[bool, float | None]:
+    def localization_of_kind(
+        generated_locations: list[Location],
+        gold_locations: list[Location],
+        kind: ScopeKind,
+    ) -> tuple[bool, float | None]:
         """
         Compute match and precision for a given scope kind.
         """
-        gold_scopes = set([scope for loc in gold_locations if (scope := loc.most_recent_scope(kind)) is not None])
-        generated_scopes = set([scope for loc in generated_locations if (scope := loc.most_recent_scope(kind)) is not None])
+        gold_scopes = set(
+            [
+                scope
+                for loc in gold_locations
+                if (scope := loc.most_recent_scope(kind)) is not None
+            ]
+        )
+        generated_scopes = set(
+            [
+                scope
+                for loc in generated_locations
+                if (scope := loc.most_recent_scope(kind)) is not None
+            ]
+        )
 
         match = all(gold_scope in generated_scopes for gold_scope in gold_scopes)
-        precision = len(gold_scopes & generated_scopes) / len(generated_scopes) if generated_scopes else None
+        precision = (
+            len(gold_scopes & generated_scopes) / len(generated_scopes)
+            if generated_scopes
+            else None
+        )
 
         return match, precision
+
+    @staticmethod
+    def from_locations(
+        generated_locations: list[Location], gold_locations: list[Location]
+    ) -> LocalizationMetrics:
+        """
+        Compute localization metrics for a list of locations.
+
+        Args:
+            generated_locations: The locations generated by the model.
+            gold_locations: The golden locations acting as ground truth.
+        """
+        # File statistics
+        file_match, file_precision = LocalizationMetrics.localization_of_kind(
+            generated_locations, gold_locations, ScopeKind.FILE
+        )
+
+        # Function statistics
+        function_match, function_precision = LocalizationMetrics.localization_of_kind(
+            generated_locations, gold_locations, ScopeKind.FUNCTION
+        )
+
+        # Class statistics
+        class_match, class_precision = LocalizationMetrics.localization_of_kind(
+            generated_locations, gold_locations, ScopeKind.CLASS
+        )
+
+        return LocalizationMetrics(
+            file_match=file_match,
+            function_match=function_match,
+            class_match=class_match,
+            file_precision=file_precision,
+            function_precision=function_precision,
+            class_precision=class_precision,
+        )
 
     @staticmethod
     def from_patch(generated_patch: Patch, gold_patch: Patch) -> LocalizationMetrics:
@@ -58,21 +115,5 @@ class LocalizationMetrics(BaseModel):
         """
         generated_locations = generated_patch.locations
         gold_locations = gold_patch.locations
-    
-        # File statistics
-        file_match, file_precision = LocalizationMetrics.localization_of_kind(generated_locations, gold_locations, ScopeKind.FILE)
 
-        # Function statistics
-        function_match, function_precision = LocalizationMetrics.localization_of_kind(generated_locations, gold_locations, ScopeKind.FUNCTION)
-
-        # Class statistics
-        class_match, class_precision = LocalizationMetrics.localization_of_kind(generated_locations, gold_locations, ScopeKind.CLASS)
-
-        return LocalizationMetrics(
-            file_match=file_match,
-            function_match=function_match,
-            class_match=class_match,
-            file_precision=file_precision,
-            function_precision=function_precision,
-            class_precision=class_precision,
-        )
+        return LocalizationMetrics.from_locations(generated_locations, gold_locations)
